Move bank cache out of resource
